Move jekyll build into federalist step
Gatsby expects to own the entire site directory. Jekyll does too, but Jekyll
will also copy over pre-built files for you. The problem is that dumping the
Gatsby build files into the source directory causes some other headaches (how do
you cleanup the generated files reliably?). Instead we want to copy the Gatsby
build files directly into the Jekyll output, `_site`. By default, Jekyll blows
away the `_site` directory before building. It's `keep_files` option won't work
with globs, and so it makes it impossible to whitelist the Gatsby build files.
So, we'd like to build Jekyll first, and then build Gatsby so that any Gatsby
files will overwrite the Jekyll files and Jekyll won't blow away the _site
directory.

When using the Jekyll build option, Federalist builds the JS first with `npm run
federalist`, and then runs Jekyll. We want to reverse that by running Jekyll
first and then run the gatsby build, however, this is not supported by
Federalist[1]. Instead, we can move the Jekyll build into the npm step manually,
but this means the Ruby/Jekyll dependencies are not set up yet and have to be
setup manually, hence the setup in `_bin/federalist-build.sh`.

Finally, once Jekyll and Gatsby are built in the npm step, Federalist will still
run Jekyll which will cause issues. So, we use a dummy configuration that sets
the Jekyll source as `_stage`, which is where we build into. Then the second
Jekyll step just copys all our built files as-is to the Federalist build
directory `_site`. Pain in the butt, but it works and encourages us to move off
of Jekyll. The nice thing is that when Jekyll goes away, all of this does too
and there are no routing hacks that have to be made to Gatsby.

[1]: https://github.com/18F/federalist/issues/1923